# docs/openapi/critical-endpoints.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Adaptive Tamil Platform Critical API Contracts",
    "version": "1.0.0",
    "description": "Critical endpoint OpenAPI contract used for launch-readiness response governance."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/lessons": {
      "get": {
        "tags": ["Lessons"],
        "operationId": "listLessons",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lesson list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/LessonEntity" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Lessons"],
        "operationId": "createLesson",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created lesson",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonEntity"
                }
              }
            }
          }
        }
      }
    },
    "/lessons/{id}": {
      "put": {
        "tags": ["Lessons"],
        "operationId": "updateLesson",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated lesson",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonEntity"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Lessons"],
        "operationId": "deleteLesson",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete confirmation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteLessonResponse"
                }
              }
            }
          }
        }
      }
    },
    "/lessons/categories": {
      "get": {
        "tags": ["Lessons"],
        "operationId": "getCategorySummary",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Category summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategorySummaryItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/lessons/stages/progress": {
      "get": {
        "tags": ["Lessons"],
        "operationId": "getStageProgress",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Progress by stage",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StageProgressItem"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LessonEntity": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "_id",
          "category",
          "difficulty",
          "type",
          "question",
          "correct_answer",
          "stage",
          "stageOrder",
          "exerciseType",
          "order"
        ],
        "properties": {
          "_id": { "type": "string" },
          "category": {
            "type": "string",
            "enum": ["uyir", "mei", "uyir-mei", "grammar", "sentences"]
          },
          "difficulty": {
            "type": "string",
            "enum": ["Beginner", "Intermediate", "Advanced"]
          },
          "type": {
            "type": "string",
            "enum": ["mcq", "text"]
          },
          "question": { "type": "string" },
          "correct_answer": { "type": "string" },
          "stage": { "type": "integer", "minimum": 1, "maximum": 10 },
          "stageOrder": { "type": "integer", "minimum": 0 },
          "exerciseType": { "type": "string" },
          "order": { "type": "integer", "minimum": 0 },
          "isMasteryTest": { "type": "boolean" },
          "unlocksStage": { "type": "integer", "minimum": 1, "maximum": 10 }
        }
      },
      "CreateLessonRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "category",
          "difficulty",
          "type",
          "question",
          "options",
          "correct_answer",
          "stage",
          "stageOrder",
          "order",
          "exerciseType"
        ],
        "properties": {
          "category": {
            "type": "string",
            "enum": ["uyir", "mei", "uyir-mei", "grammar", "sentences"]
          },
          "difficulty": {
            "type": "string",
            "enum": ["Beginner", "Intermediate", "Advanced"]
          },
          "type": {
            "type": "string",
            "enum": ["mcq", "text"]
          },
          "question": { "type": "string", "minLength": 1 },
          "options": {
            "type": "array",
            "minItems": 2,
            "items": { "type": "string" }
          },
          "correct_answer": { "type": "string", "minLength": 1 },
          "stage": { "type": "integer", "minimum": 1, "maximum": 10 },
          "stageOrder": { "type": "integer", "minimum": 0 },
          "order": { "type": "integer", "minimum": 0 },
          "exerciseType": { "type": "string" }
        }
      },
      "DeleteLessonResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["message"],
        "properties": {
          "message": { "type": "string" }
        }
      },
      "CategorySummaryItem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["_id", "count"],
        "properties": {
          "_id": {
            "type": "object",
            "additionalProperties": false,
            "required": ["category", "difficulty"],
            "properties": {
              "category": { "type": "string" },
              "difficulty": { "type": "string" }
            }
          },
          "count": { "type": "integer", "minimum": 0 }
        }
      },
      "StageProgressItem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["stage", "totalLessons", "completedLessons", "unlocked", "masteryPassed"],
        "properties": {
          "stage": { "type": "integer", "minimum": 1, "maximum": 10 },
          "totalLessons": { "type": "integer", "minimum": 0 },
          "completedLessons": { "type": "integer", "minimum": 0 },
          "unlocked": { "type": "boolean" },
          "masteryPassed": { "type": "boolean" }
        }
      },
...
    }
  }
}